update the bucket name, object name's special issue for all the others apis.

+++ KS3SDKIOS/KS3YunSDK/KS3YunSDK/KS3YunSDK/KSYCore.RunTime/KS3ServiceRequest.m
@@ -50,14 +50,11 @@
 }
 
 - (NSString *)URLEncodedString:(NSString *)str
-
 {
-    
     NSMutableString *output = [NSMutableString string];
-    
     const unsigned char *source = (const unsigned char *)[str UTF8String];
     
-    int sourceLen = strlen((const char *)source);
+    int sourceLen = (int)strlen((const char *)source);
     
     for (int i = 0; i < sourceLen; ++i) {
